Correccion total

+++ src/controllers/postController.js
@@ -1,22 +1,66 @@
-const { Post } = require('../models');
-const { Post_Image } = require('../models');
-const { User } = require('../models');
-const { Tag } = require('../models');
+const { Post, Post_Image, User, Tag, Comment } = require('../models');
 
-const getPosts = async (req, res) => {
-    const data = await Post.find()
-        .populate("nickName")
-        .populate("imagenes")
-        .populate("tags");
-    res.status(200).json(data);
+const getPosts = async (_, res) => {
+    try {
+        const data = await Post.find()
+            .populate({
+                path: 'user',
+                select: ['nickName', 'email']
+            })
+            .populate({
+                path: 'imagenes',
+                select: 'url'
+            })
+            .populate({
+                path: 'tags',
+                select: 'tagName'
+            })
+        res.status(200).json(data);
+
+    } catch (error) {
+        res.status(500).json({ error: error.message });
+    }
 };
 
 const getPostById = async (req, res) => {
-    const data = await Post.findById(req.params.id)
-        .populate("imagenes")
-        .populate("tags");
-    if (!data) return res.status(404).json({ error: 'Post No Encontrado' });
-    res.status(200).json(data);
+    try {
+        const idPost = req.params.id
+        const post = await Post.findById(idPost)
+            .populate({
+                path: 'user',
+                select: ['nickName', 'email']
+            })
+            .populate({
+                path: 'imagenes',
+                select: 'url'
+            })
+            .populate({
+                path: 'tags',
+                select: 'tagName'
+            })
+
+        if (!post) return res.status(404).json({ error: 'Post No Encontrado' });
+
+        const seisMesesAtras = new Date();
+        seisMesesAtras.setMonth(seisMesesAtras.getMonth() - 6);
+        const comments = await Comment.find({
+            $and: [
+                { postId: idPost },
+                { fecha: { $gte: seisMesesAtras } }]
+        })
+            .populate({
+                path: 'postId',
+                select: '_id'
+            })
+            .populate({
+                path: 'user',
+                select: 'nickName'
+            })
+
+        res.status(200).json({ post, comments });
+    } catch (error) {
+        res.status(500).json({ error: error.message });
+    }
 };
 
 const createPost = async (req, res) => {
@@ -26,7 +70,6 @@ const createPost = async (req, res) => {
         if (!user) {
             return res.status(404).json({ message: `No existe ningun usuario con el nickName ${name}` });
         }
-        //const newPost = await Post.create({ description: req.body.description, nickName: name, fecha: req.body.fecha, user: user._id });
         let imagenIds = [];
 
         if (req.body.imagenes && Array.isArray(req.body.imagenes)) {
@@ -51,13 +94,14 @@ const createPost = async (req, res) => {
             }
         }
 
-        const newPost = await Post.create({
+        const newPost = new Post({
             description: req.body.description,
-            nickName: user._id,
+            user: user._id,
             fecha: req.body.fecha,
             imagenes: imagenIds,
             tags: tagIds
         });
+        await newPost.save()
         res.status(201).json(newPost);
     } catch (error) {
         res.status(400).json({ error: error.message });
@@ -65,42 +109,60 @@ const createPost = async (req, res) => {
 };
 
 const deletePostById = async (req, res) => {
-    const data = await Post.findById(req.params.id)
-    if (!data) return res.status(404).json({ error: 'Post No Encontrado' });
-    const removed = await data.deleteOne();
-    res.status(200).json(removed);
+    try {
+        const idPost = req.params.id
+        const post = await Post.findByIdAndDelete(idPost)
+        if (!post) return res.status(404).json({ error: 'Post No Encontrado' });
+        await Comment.deleteMany({ postId: idPost })
+        res.status(200).json(post);
+    }
+    catch (error) {
+        res.status(400).json({ error: error.message });
+    }
 };
-
-//Falta probar de aca para abajo ***********************************
 
 const putPostById = async (req, res) => {
-    await Post.updateOne({ _id: req.params.id }, req.body);
-    const data = await Post.findById(req.params.id);
-    if (!data) return res.status(404).json({ error: 'Post No Encontrado' });
-    res.status(201).json(data);
+    try {
+        const idPost = req.params.id
+        const udPost = await Post.findByIdAndUpdate(
+            { _id: idPost },
+            req.body,
+            { new: true }
+        );
+        if (!udPost) return res.status(404).json({ error: 'Post No Encontrado' });
+        res.status(201).json(udPost);
+    }
+    catch (error) {
+        res.status(400).json({ error: error.message });
+    }
 };
 
-const postPostImageById = async (req, res) => {
-    const { id } = req.params;
-    const { url } = req.body;
+const putPostImage = async (req, res) => {
+    try {
+        const idPost = req.params.id;
+        const url = req.body.url;
 
-    const post = await Post.findById(id);
-    if (!post) return res.status(404).json({ error: "Post no encontrado" });
+        const post = await Post.findById(idPost);
+        if (!post) return res.status(404).json({ error: "Post no encontrado" });
 
-    const newImage = await Post_Image.create({ url, postId: id });
+        const newImage = await Post_Image.create({ url, postId: idPost });
 
-    post.imagenes.push(newImage._id);
-    await post.save();
+        post.imagenes.push(newImage._id);
+        await post.save();
 
-    res.status(201).json(newImage);
+        res.status(200).json(newImage);
+    }
+    catch (error) {
+        res.status(400).json({ error: error.message });
+    }
 };
 
 
 const putPostImageById = async (req, res) => {
-    const { id, imageId } = req.params;
+    const { idPost, imageId } = req.params;
     const { url } = req.body;
 
-    const image = await Post_Image.findOne({ _id: imageId});
+    const image = await Post_Image.findOne({ _id: imageId });
     if (!image) return res.status(404).json({ error: "Imagen no encontrada" });
 
     image.url = url;
@@ -111,10 +173,10 @@ const putPostImageById = async (req, res) => {
 
 const deletePostImgById = async (req, res) => {
     const { id, imageId } = req.params;
-
-    const image = await Post_Image.findOne({ _id: imageId});
+    const post = await Post.findOne({ _id: id });
+    const image = await Post_Image.findOne({ _id: imageId });
     if (!image) return res.status(404).json({ error: "Imagen no encontrada" });
-
+    if (!post) return res.status(404).json({ error: "Post no encontrada" });
     await image.deleteOne();
 
     res.status(200).json({ message: "Imagen eliminada exitosamente" });
@@ -128,5 +190,5 @@ module.exports = {
     putPostImageById,
     putPostById,
     deletePostImgById,
-    postPostImageById
+    putPostImage
 };
